% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eat.R
\name{eat}
\alias{eat}
\title{Eat columns from another data frame}
\usage{
eat(x, y, ..., by = NULL, fun = NULL, check = "~j",
  conflicts = c("keep_xy", "keep_x", "keep_y", "patch"), prefix = NULL)
}
\arguments{
\item{x}{tbls to join}

\item{y}{tbls to join}

\item{...}{One or more unquoted expressions, passed to \code{dplyr::select},
defining the columns to keep from \code{y}.}

\item{by}{a character vector of variables to join by.  If \code{NULL}, the
default, \code{*_join()} will do a natural join, using all variables with
common names across the two tables. A message lists the variables so
that you can check they're right (to suppress the message, simply
explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector.
For example, \code{by = c("a" = "b")} will match \code{x.a} to
\code{y.b}.}

\item{fun}{function or formula or \code{NULL}, if not \code{NULL}, \code{y}
will be grouped by its \code{by} columns and \code{fun} will be applied to
all kept columns from {y}.}

\item{check}{string which determines what will trigger a message/warning or
a failure, see details}

\item{prefix}{prefix of new columns or function to apply on names of new
columns, supports formula notation to reformat the new names further if
needed}

\item{replace}{in case of column conflict, should we keep both (default)
or replace/ditch one of them ?}
}
\value{
a data frame
}
\description{
Modified left_join where only a specified subset of \code{y} is kept, with
  optional checks and transformations.
}
