% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_matches.R
\name{match_equal}
\alias{match_equal}
\title{Title}
\usage{
match_equal(
  lhs,
  rhs,
  ignore_case = FALSE,
  ignore_element_order = FALSE,
  ignore_row_order = FALSE,
  ignore_class = FALSE
)
}
\arguments{
\item{lhs}{expression giving the values to match on the lhs}

\item{rhs}{expression giving the values to match on the rhs}

\item{ignore_case}{boolean, if \code{TRUE} not only will character columns be set
to lower case before matching, but \code{rapply} will be used on lists and data
frame to make all characters lower case before comparison}

\item{ignore_element_order}{boolean, if \code{TRUE} elements of lists (including
data frames) will be sorted before comparison}

\item{ignore_row_order}{boolean, if \code{TRUE} data frames will be sorted before
comparison}

\item{ignore_class}{boolean, if \code{TRUE} unclass will be used on elements before
comparison, so for instance tibbles and lists can be}
}
\description{
Title
}
