% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safejoin-package.R
\docType{package}
\name{safejoin-package}
\alias{safejoin}
\alias{safejoin-package}
\title{safejoin: Perform "Safe" Table Joins}
\description{
The goal of 'safejoin' is to guarantee that when performing joins extra rows are not added to your data. 'safejoin' provides a wrapper around 'dplyr::left_join' that will raise an error when extra rows are unexpectedly added to your data. This can be useful when working with data where you expect there to be a many to one relationship but you are not certain the relationship holds.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SamEdwardes/safejoin}
  \item Report bugs at \url{https://github.com/SamEdwardes/safejoin/issues}
}

}
\author{
\strong{Maintainer}: Sam Edwardes \email{edwardes.s@gmail.com}

}
\keyword{internal}
