% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_fuzzy.R
\name{XY}
\alias{XY}
\alias{X}
\alias{Y}
\title{Select a variable from to opperate a fuzzy join}
\usage{
X(var)

Y(var)
}
\arguments{
\item{var}{a string containing the name of the variable to select}
}
\description{
#' Should be called only inside a formula fed to the match_fun argument of a
safejoin function.
}
\details{
Use \code{X} to select a variable from the lhs, and \code{Y} to select a variable from
the rhs, calls to \code{X} and \code{Y} will be detected to register the variables used
so that the proper cartesian product can be operated.
}
