% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safejoin-package.R
\docType{package}
\name{safejoin-package}
\alias{safejoin}
\alias{safejoin-package}
\title{safejoin: Join safely and Deal with Conflicting Columns}
\description{
The package *safejoin* features wrappers around packages *dplyr* and
    *fuzzyjoin*'s functions to join safely using various checks. It also comes
    packed with features to select columns, rename them, operate on conflicting 
    ones (coalesce for example), or aggregate the rhs on the joining columns
    before joining.
}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

}
\keyword{internal}
